MAINT: remove special case for scalar arrays indexed by bools
The branch was only necessary to support this while
`PyArray_PyIntAsIntp` allowed for boolean to be converted to ints.

There is a slight thing here that array-likes may behave slightly
difference from numpy arrays (no-copy on 0-d) and if they did not
deprecate `operator.index(non_0d_array_like)` things get strange,
but that problem always existed and there is nothing much we can
do about it I think.