WIP: Dirty way to accept dict views
This is probably not right, but shows that it is OK to accept them
in principle. We may want to accept only Sequence, in which
case python would have to register it.

This also cannot be done on older pythons where the fixed order is
not a language feature (also fill in that this is actually a language
feature).