WIP,API: Channel at least some of the DType transfer through new API.
Right now it is very unclear how to design this, the only good option is
probably to cram most things into a CastingImpl and for now only provide
a very limited API to create new CastingImpl objects to downstream.

This is very much lacking that we should look up casting functions from
the castingimpl, and never go back to the PyArray_Descr!