Register floats (need replacement with abstract) and remove custom error
The custom error just slows down things a lot ;). 50% slower, mostly
due to the added cost of checking whether casting is possible, which
is expensive, because many ufuncs will try many times to find the
correct loop.
That should not matter, because I wish to cache/map type signature
to UFuncImpl in the end.
Furthermore, the fallback casting, creates both dtype instances
and the CastingImpl object unnecessarily, so will be much faster
after a refactor in any case.