UFUNC: First version which works somewhat
IPython starts, ufunc loop gets resolved with array input. Right
now the array coercion (at least if we already have an array) is
a bit sluggish, a good chunk due to the coercion cache object
allocation apparently... We run at about 1.2 the time of the old
code in this case. That seems not awfully bad, but considering
there are some other speed enhancements already hidden in here,
not ideal either.