UFUNC: Add resolvers in python and fix logic
The current logic gives precedence to abstract dtypes/None (without
a hierarchy) as opposed to concrete ones for not-given (out)
DTypes.
This is so that `OO->O` and `OO->?` can be registered by an
`OO->AnyDType` resolver can fix the ambiguity.

The current state segfaults in some garbage collection work, so
more refcounting fixes are probably in order next (and checking
whether the the GC tracking is implemented correctly).