ENH: Allow zero iterator to allocate when all operands
Previously when all operands were allocated/passed in as NULL/None,
the iterator would reject to allocate the outputs.
This now allows the iterator to allocate the output. The output
will be 0-D when no

The fact that this was disallowed probably is a bit more obvious
when realizing that the

This will allow gufuncs with a signature such as `->(3,3)` (i.e.
only outputs). However, gufuncs and ufuncs in general may still
have issues with the output dtype or generally the correct loop
discovery which is tied to the inputs (if there is more than one
loop).