Continue with dispatching
This needs some thoughts. Mainly around dispatching also on outputs
(which generally should not be necessary, but maybe doesn't matter).

Also, I am turning and tossing the boundarraymethod idea a bit, wondering
if pushing the "reference problem" into the ufunc is the right approach
(I think it probably is, since it seems very reasonable for simple cases
like casting, but not quite sure.)