Implement (untested) promotion using a python function
I decided to limit it to returning a new type tuple. The user can
register a new loop in the promotion function itself!

I am not considering making a PyCapsule mechanism to allow for low level
C implementation of the promotion. That would make the promotion much
quicker.
It would be simple enough, but only worthwhile if I can push the
value-based promotion into the actual promotion function (which may
well work!).  In that case, it would be worthwhile, since the promotion
will be called almost always when there is 0-D input. Otherwise, it does
not really seem worthwhile.