MAINT: Refactor UFunc core to use NEP 43 style dispatching/promotion
This PR replaces the dispaching, promotion and promotion with those
discussed in NEP 43. The is, the inner-loop is represented by
an ArrayMethod, promotion can either be legacy (fallback to to previous)
or a function (not yet used).

Dispatching further uses a cache of promotion results based on the
input and user provided DTypes.
This cache is currently implemented as light-weight custom hashmap.

As of this commit, it is expected that some details will be in flux,
e.g. how exactly the cache is setup (e.g. it could have a maximum size
or grow on "exessive" cache misses instead of load factor).

The hashmap is a new file at: numpy/core/src/common/npy_hashtable.c

The dispatching and promotion is implemented in:
numpy/core/src/umath/dispatching.c

The current ufunc loops are unmodified, but must be wrapped into
the new ArrayMethod, this happens in:
numpy/core/src/umath/legacy_array_method.c

---

Further changes are cleanups, mainly enabled by bringing the masked
and non-masked loops closer together, as well as refactors necessary
to weave the new dispatched ArrayMethods as `ufuncimpl` into the
ufunc machinery.

To keep this changeset a bit smaller, it only modifies ufuncs and
generalized ufuncs, but not reduction(like) or `ufunc.at`.