MAINT: Refactor UFunc core to use NEP 43 style dispatching/promotion
This PR replaces the dispaching, promotion and promotion with those
discussed in NEP 43. The is, the inner-loop is represented by
an ArrayMethod, promotion can either be legacy (fallback to to previous)
or a function (not yet used).

Dispatching further uses a cache of promotion results based on the
input and user provided DTypes.
This cache is currently implemented as light-weight custom hashmap.

As of this commit, it is expected that some details will be in flux,
e.g. how exactly the cache is setup (e.g. it could have a maximum size
or grow on "exessive" cache misses instead of load factor).

The hashmap is a new file at: numpy/core/src/common/npy_hashtable.c

The dispatching and promotion is implemented in:
numpy/core/src/umath/dispatching.c

The current ufunc loops are unmodified, but must be wrapped into
the new ArrayMethod, this happens in:
numpy/core/src/umath/legacy_array_method.c

---

Further changes are cleanups, mainly enabled by bringing the masked
and non-masked loops closer together, as well as refactors necessary
to weave the new dispatched ArrayMethods as `ufuncimpl` into the
ufunc machinery.

To keep this changeset a bit smaller, it only modifies ufuncs and
generalized ufuncs, but not reduction(like) or `ufunc.at`.

MAINT: Use new promotion/dispatching for reductions

Add logical promoters and special case...

... but all the paths in reduceat, etc. need to be adapted now
until the _real_ testing can start

Finish refactor of the normal reduction loop

Everything seems to work now (probably there are still some issues,
since the accumulate/reduceat probably don't work, I did not try
to run the testsuit yet)

Convert accumulate to new system

Accumulate still has a few quirks remaining probably, but for now we need
to convert it, not try to think through all the quirks...

Give reduceat the accumulate treatment (tests not run yet)

Undo changes to out refcounting

The move was just incorrect for the nditer path...

BUG: Fix inverted GIL release

BUG: Do not cache promoters...

If promoters only modify the DTypes, we can't cache them reasonably

Fix registereing all legacy loops explicitly.

This is necessary for the logical ones, since they stop promoting
using the legacy mechanism (they will always resolve with the
new one successfully).

BUG: Fix "too tired to code straight typo"...

Further fixes, and small changes, should "backport" some refactoring probably

Use unsafe casting for now in reducelikes (at least reductions had that)

MAINT,TST: Necessary fixes to tests and MA to make everything pass

BUg: do not check casting on custom resolutoin

This might override casting (e.g. in astropy/pyerfa) and there is just
no need.  So skip it and fix pyerfa on the way.

fixups and new tests