MAINT: Fix floating point warning flag as much as possible
This PR avoids setting FPE flag flags rather than clearing them
where possible.

The main place where this is not easy is SSE (I could not yet find
the `isless` equivalent for basic SSE), and did not want to modify
the code too much.

This also updates the docs to copy the informational section to the
"Floating point error handling" landing page.  Right now, the docs
assume that we (for now) keep the behaviour of NOT giving floating
point warnings for comparisons with NaN.

Hopefully, using these functions should not have any speed impact.
There is some chance that compilers will not honor the contract
and replace `isless` with instructions that do set floating point
error flags.  If/where this happens, we may have to undo the changes.

This should have no impact on functionality (except if compilers
do not follow C99/IEEE correctly, which is entirely possible).

However, signalling NaNs WILL now warn more often (as per C99/IEEE),
since we do not clear the warnings as agressively.