DOC: Update reference to pass refguide
This makes things pass even though there should be a massive amount
of types added in other places, so that it is not complete.

But I think this may be useful, since the complete change would be
quite verbose...

One thing to note may be that `np.float64(nan)` currently does not
round-trip correctly of course, this would require adding `np.`
to the repr (or quotes).