WIP: Check last commit message
The idea here, is that we can use size 1 shape in which case more
logic is necessary.  The additional logic depends mainly on whether
the operand is allowed to be implicitly broadcast.
(As opposed to explicit broadcasting, which we always allow and
where we should be able to assume that marking as reduction is
always OK.)

When implicit broadcasting happens and that broadcast would result
in a reduction, we will have to clean it up later-on.  The logic
for that exists here, but currently kicks in when it should not
need to.