This is a rough proof of concept that breaking ABI is possible
IT does a few things:

1. We define symbols at compile time, this needs to be tweaked but
   allows us the user to opt-in to use new API.
   For functions, this is convenient, by just adding it to the def.

2. We just create a new table in the header and inject it into the
   old one.
   We will have to create the table translation once (which does
   not have to be particularly legible).

API that is not in NumPy 2, cannot be used in NumPy 2 and if you want
to support NumPy 2, you have to avoid it also on earlier versions.

We can inject new symbols (this is annoying, as it adds the functions
into a random user module, alternatively, we wrap the function into
a compat shim, that is a bit annoying, since we don't have a direct
way to spell it, but should be OK).