MAINT,TST: Fix issues due to out-of-range integers
This contains two fixes:
1. a somewhat hackish solution to keep allowing very large integers
   for unary ufuncs.  This is partially wrong, but keeps isnan and
   other functions working (reducing fallout in the tests quite a bit)
2. Generally adapt changes that now fail for the same reason, this is
   mainly due to comparisons like `-1 < uint8` now failing in generally.
   That is fixable in principle, but it isn't super easy.

Neither is perfect, both are the pragmatic solutions at this time until
they proof to be too problematic for some reason.

(Getting rid of going to object for large integers is tougher and needs
deeper work.  It certainly is possible but too difficult at this stage:
we need to remove some old branches first.
We could, and maybe should, get rid of the `uint` step though, it is even
more of a trap I think!)