WIP: Make it so we only have Int8, Int16... and no C version DTypes
This retains the python descriptor instances.  The main problem with
this approach is that it confuses the legacy loop finder, since
that searches for exact matches.

That can be worked around, by forcing the singleton (which should be
fine, only affects execution speed very mildly for metadata probably).

That part needs a bit tweaking, otherwise I added a loop using "q"
for unsigned int and signed int comparisons.  "q" isn't always the
canonical name for int64, so that loop cannot be found.