ENH: Specialize and simplify buffered iterator advancing
The old code alway went through `goto_iterindex` which needs to
inspect all dimensions with divisions.

This adds a custom advance function similar to the normal advance
(but it needs to take the outersize and outerdim into account).
This means that we often only need to touch the single `outerdim`
to advance pointers and don't need any expensive divisions (except
to figure out the outerdim size once).